refactor XmlExtract.hs to return XML instead Text
Since there is markup information that can be embedded in subtitles it
is better to explicitly show this in the type. An alternative is to
define a type just for the reduced markup capabilities of SRT. Maybe
later, iteratively.

Also add LICENSE header and merge executables since it seems I like it
better with a single binary as output.